+++ www/Library/content_samples/Image/Image.php
@@ -6,9 +6,15 @@
  */
 namespace Library\content_samples\Image;
 
-use Boolive\data\Entity;
+use Boolive\data\Entity,
+    Boolive\values\Rule;
 
 class Image extends Entity
 {
-// @todo В правиле на файл указать расширения картинок
+    public function defineRule()
+    {
+        parent::defineRule();
+        // Ассоциация с файлами с расширением css
+        $this->_rule->arrays[0]['file']->arrays[0]['name']->ospatterns('*.png', '*.jpg', '*.gif')->required();
+    }
 }