+++ www/Library/views/DirectHandler/DirectHandler.php
@@ -17,8 +17,8 @@ class DirectHandler extends View
     {
         return Rule::arrays(array(
             'REQUEST' => Rule::arrays(array(
-                'method' => Rule::is('POST'),
-                'view' => Rule::entity()->required(),
+                'method' => Rule::eq('POST'),
+                'direct' => Rule::entity()->required(),
                 )
             ))
         );
@@ -27,7 +27,7 @@ class DirectHandler extends View
     public function work()
     {
         $v = array();
-        $v['out'] = $this->_input['REQUEST']['view']->start($this->_commands, $this->_input_child);
+        $v['out'] = $this->_input['REQUEST']['direct']->start($this->_commands, $this->_input_child);
 		// Вместе с html нужно передать пути на файлы css и js
         foreach ($this->_commands->get('addHtml') as $com){
             if (in_array($com[0], array("link", "script"/*, "meta", "title", "style"*/))){