+++ www/Library/views/View/View.php
@@ -21,18 +21,18 @@ class View extends Entity
      * В качестве правила по умолчанию используется $this->getInputRule()
      * @var mixed
      */
-    //protected $_input;
+    protected $_input;
     /**
      * Ошибки при проверки входящих данных
      * @var \Boolive\errors\Error
      */
-    //protected $_input_error;
+    protected $_input_error;
     /**
      * Команды, передающиеся по всем исполняемым объектам.
      * Инициализируется в методе start()
      * @var \Boolive\commands\Commands
      */
-    //protected $_commands;
+    protected $_commands;
     /**
      * Входящие данные для подчиенных объектов
      * @var mixed
@@ -87,7 +87,7 @@ class View extends Entity
             ob_start();
                 // Выполнение своей работы
                 $result = $this->work();
-                if ($result === false){
+                if ($result === false || is_array($result)){
                     return $result;
                 }else{
                     $result = ob_get_contents().$result;